contact policy

+++ app/Http/Controllers/ContactController.php
@@ -4,6 +4,7 @@ namespace App\Http\Controllers;
 
 
 use App\Models\Contact;
+use Illuminate\Auth\Access\Gate;
 use Illuminate\Http\Request;
 
 class ContactController extends Controller
@@ -61,6 +62,8 @@ class ContactController extends Controller
      */
     public function show(Contact $contact)
     {
+        $this->authorize('view', $contact);
+
         return view('contacts.show', compact('contact'));
     }
 
@@ -73,6 +76,8 @@ class ContactController extends Controller
      */
     public function edit(Contact $contact)
     {
+        $this->authorize('update', $contact);
+
         return view('contacts.edit', compact('contact'));
     }
 
@@ -86,6 +91,8 @@ class ContactController extends Controller
      */
     public function update(Request $request, Contact $contact)
     {
+        $this->authorize('update', $contact);
+
         $data = $request->validate([
             'name' => 'required',
             'phone_number' => 'required|digits:9',
@@ -107,6 +114,8 @@ class ContactController extends Controller
      */
     public function destroy(Contact $contact)
     {
+        $this->authorize('delete', $contact); # chequeo que usuario puede eliminar solo si es SU contacto
+
         $contact->delete();
         
         return redirect()->route('home');
